# openapi_schema.yaml
openapi: 3.1.0
info:
  title: China Securities Data Proxy
  description: >-
    REST API-обёртка над AKShare для доступа к финансовой отчётности,
    дивидендам, дневным котировкам и валютным курсам China A-share (SSE/SZSE)
    и Hong Kong (HKEX) тикеров.
  version: 1.0.0
servers:
  - url: https://china-securities-proxy.onrender.com
paths:
...
  /forecast/em/{symbol}:
    get:
      operationId: getForecastEastmoney
      summary: >-
        Консенсус-прогноз аналитиков Eastmoney (东财) по A-share тикеру.
        Только A-share, код без биржевого суффикса (напр. "601939").
      parameters:
        - name: symbol
          in: path
          required: true
          description: >-
            Код тикера без биржевого суффикса. Для A-share — 6 цифр (напр.
            "601939"). Для HK — код без .HK, ведущие нули не обязательны
            (напр. "939" эквивалентно "00939").
          schema:
            type: string
        - name: include_past
          in: query
          required: false
          description: >-
            По умолчанию false — возвращаются только прогнозы на текущий год
            и последующие, уже прошедшие годы отфильтровываются. true — вернуть
            всё, включая устаревшие периоды; запрашивай только если пользователь
            явно просит исторические прогнозы.
          schema:
            type: boolean
            default: false
      responses:
        "200":
          description: >-
            Прогноз консенсуса аналитиков либо found:false, если по данному
            тикеру нет покрытия у этого источника
          content:
            application/json:
              schema:
                type: object
                properties:
                  source:
                    type: string
                  symbol:
                    type: string
                  today:
                    type: string
                    description: >-
                      Сегодняшняя дата по часам сервера (YYYY-MM-DD). Считай
                      её текущей датой, а не ту, что модель помнит сама.
                  retrieved_at:
                    type: string
                    description: >-
                      Момент получения данных из источника (ISO 8601, UTC).
                      Цитируй его как дату данных; никогда не подставляй
                      прогнозные цифры по памяти модели.
                  dropped_past_periods:
                    type: array
                    description: >-
                      Периоды, отброшенные как уже прошедшие (по умолчанию
                      возвращаются только прогнозы на текущий год и позже).
                    items:
                      type: string
                  found:
                    type: boolean
                  note:
                    type: string
                  data:
                    type: array
                    items:
                      type: object
        "502":
          description: Ошибка при обращении к источнику данных (AKShare)
          content:
            application/json:
              schema:
                type: object
                properties:
                  detail:
                    type: string
  /forecast/ths/{symbol}:
    get:
      operationId: getForecastThs
      summary: >-
        Консенсус-прогноз аналитиков Tonghuashun/同花顺 по A-share тикеру.
        Только A-share, код без биржевого суффикса (напр. "601939").
      parameters:
        - name: symbol
          in: path
          required: true
          description: >-
            Код тикера без биржевого суффикса. Для A-share — 6 цифр (напр.
            "601939"). Для HK — код без .HK, ведущие нули не обязательны
            (напр. "939" эквивалентно "00939").
          schema:
            type: string
        - name: indicator
          in: query
          required: false
          description: >-
            Раздел прогноза THS. Один из: "预测年报每股收益",
            "预测年报净利润", "业绩预测详表-机构", "业绩预测详表-详细指标预测"
            (по умолчанию). Значение должно совпадать ТОЧНО, иначе будет
            пустой результат.
          schema:
            type: string
            default: 业绩预测详表-详细指标预测
        - name: include_past
          in: query
          required: false
          description: >-
            По умолчанию false — возвращаются только прогнозы на текущий год
            и последующие, уже прошедшие годы отфильтровываются. true — вернуть
            всё, включая устаревшие периоды; запрашивай только если пользователь
            явно просит исторические прогнозы.
          schema:
            type: boolean
            default: false
      responses:
        "200":
          description: >-
            Прогноз консенсуса аналитиков либо found:false, если по данному
            тикеру нет покрытия у этого источника
          content:
            application/json:
              schema:
                type: object
                properties:
                  source:
                    type: string
                  symbol:
                    type: string
                  today:
                    type: string
                    description: >-
                      Сегодняшняя дата по часам сервера (YYYY-MM-DD). Считай
                      её текущей датой, а не ту, что модель помнит сама.
                  retrieved_at:
                    type: string
                    description: >-
                      Момент получения данных из источника (ISO 8601, UTC).
                      Цитируй его как дату данных; никогда не подставляй
                      прогнозные цифры по памяти модели.
                  dropped_past_periods:
                    type: array
                    description: >-
                      Периоды, отброшенные как уже прошедшие (по умолчанию
                      возвращаются только прогнозы на текущий год и позже).
                    items:
                      type: string
                  found:
                    type: boolean
                  note:
                    type: string
                  data:
                    type: array
                    items:
                      type: object
        "400":
          description: Неверное значение indicator
          content:
            application/json:
              schema:
                type: object
                properties:
                  detail:
                    type: string
        "502":
          description: Ошибка при обращении к источнику данных (AKShare)
          content:
            application/json:
              schema:
                type: object
                properties:
                  detail:
                    type: string
  /forecast/hk/{symbol}:
    get:
      operationId: getForecastHk
      summary: >-
        Консенсус-прогноз аналитиков ET Net/经济通 по гонконгскому тикеру
        (основной источник прогнозов для HK-бумаг, напр. BOC/CCB/CITIC).
        Код без суффикса .HK, ведущие нули не обязательны (напр. "939",
        "3988", "267").
      parameters:
        - name: symbol
          in: path
          required: true
          description: >-
            Код тикера без биржевого суффикса. Для A-share — 6 цифр (напр.
            "601939"). Для HK — код без .HK, ведущие нули не обязательны
            (напр. "939" эквивалентно "00939").
          schema:
            type: string
        - name: indicator
          in: query
          required: false
          description: >-
            Раздел прогноза ET Net. Один из: "评级总览",
            "去年度业绩表现", "综合盈利预测" (по умолчанию), "盈利预测概览".
          schema:
            type: string
            default: 综合盈利预测
        - name: include_past
          in: query
          required: false
          description: >-
            По умолчанию false — возвращаются только прогнозы на текущий год
            и последующие, уже прошедшие годы отфильтровываются. true — вернуть
            всё, включая устаревшие периоды; запрашивай только если пользователь
            явно просит исторические прогнозы.
          schema:
            type: boolean
            default: false
      responses:
        "200":
          description: >-
            Прогноз консенсуса аналитиков либо found:false, если по данному
            тикеру нет покрытия у этого источника
          content:
            application/json:
              schema:
                type: object
                properties:
                  source:
                    type: string
                  symbol:
                    type: string
                  today:
                    type: string
                    description: >-
                      Сегодняшняя дата по часам сервера (YYYY-MM-DD). Считай
                      её текущей датой, а не ту, что модель помнит сама.
                  retrieved_at:
                    type: string
                    description: >-
                      Момент получения данных из источника (ISO 8601, UTC).
                      Цитируй его как дату данных; никогда не подставляй
                      прогнозные цифры по памяти модели.
                  dropped_past_periods:
                    type: array
                    description: >-
                      Периоды, отброшенные как уже прошедшие (по умолчанию
                      возвращаются только прогнозы на текущий год и позже).
                    items:
                      type: string
                  found:
                    type: boolean
                  note:
                    type: string
                  data:
                    type: array
                    items:
                      type: object
        "400":
          description: Неверное значение indicator
          content:
            application/json:
              schema:
                type: object
                properties:
                  detail:
                    type: string
        "502":
          description: Ошибка при обращении к источнику данных (AKShare)
          content:
            application/json:
              schema:
                type: object
                properties:
                  detail:
                    type: string
...
  /forecast/aggregate/{symbol}:
    get:
      operationId: getForecastAggregate
      summary: >-
        Собирает все применимые источники прогноза консенсуса по одному
        тикеру за один вызов: Eastmoney и Tonghuashun — если symbol похож
        на A-share (6 цифр); ET Net — если передан hk_symbol, либо сам
        symbol гонконгский.
      parameters:
        - name: symbol
          in: path
          required: true
          description: >-
            Код тикера без биржевого суффикса. Для A-share — 6 цифр (напр.
            "601939"). Для HK — код без .HK, ведущие нули не обязательны
            (напр. "939" эквивалентно "00939").
          schema:
            type: string
        - name: hk_symbol
          in: query
          required: false
          description: >-
            Гонконгский код тикера, если он отличается от symbol (напр.
            когда symbol — это A-share код, а нужен ещё и HK-прогноз по
            двойному листингу A+H).
          schema:
            type: string
        - name: include_past
          in: query
          required: false
          description: >-
            По умолчанию false — возвращаются только прогнозы на текущий год
            и последующие, уже прошедшие годы отфильтровываются. true — вернуть
            всё, включая устаревшие периоды; запрашивай только если пользователь
            явно просит исторические прогнозы.
          schema:
            type: boolean
            default: false
      responses:
        "200":
          description: >-
            Объект с ключом sources, где для каждого применимого источника
            (eastmoney / 10jqka / etnet) — свой ответ в формате как у
            соответствующего одиночного эндпоинта, либо {"error": "..."}
          content:
            application/json:
              schema:
                type: object
                properties:
                  symbol:
                    type: string
                  today:
                    type: string
                    description: >-
                      Сегодняшняя дата по часам сервера (YYYY-MM-DD). Считай
                      её текущей датой, а не ту, что модель помнит сама.
                  retrieved_at:
                    type: string
                  sources:
                    type: object
        "400":
          description: Не удалось определить ни один применимый источник
          content:
            application/json:
              schema:
                type: object
                properties:
                  detail:
                    type: string
...